Refactor package structure for app modules
Moved and renamed several classes to better organize the codebase by feature, including splitting 'utils', 'editor', and 'config' classes into 'app', 'project', 'editor/explorer', 'editor/statusbar', 'editor/terminal', and 'ui/splash' packages. Updated imports and references throughout the code to reflect new locations and class names. This improves maintainability and clarity of module boundaries.
