+++ src/main/java/pseudopad/editor/explorer/FileTreeCellRenderer.java
