+++ src/main/java/pseudopad/editor/statusbar/MemoryUsageWidget.java
@@ -1,4 +1,4 @@
-package pseudopad.editor;
+package pseudopad.editor.statusbar;
 
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
