+++ src/main/java/pseudopad/editor/terminal/TerminalPane.java
@@ -1,4 +1,4 @@
-package pseudopad.editor;
+package pseudopad.editor.terminal;
 
 import java.awt.Color;
 import java.awt.Font;
@@ -10,8 +10,6 @@ import javax.swing.text.BadLocationException;
 import javax.swing.text.SimpleAttributeSet;
 import javax.swing.text.StyleConstants;
 import javax.swing.text.StyledDocument;
-
-import pseudopad.editor.terminal.TerminalBackend;
 
 /**
  * A TextPane that acts as a terminal emulator.
