+++ src/main/java/pseudopad/project/ProjectManager.java
@@ -1,11 +1,11 @@
-package pseudopad.utils;
+package pseudopad.project;
 
 import com.google.gson.Gson;
+
 import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
-import pseudopad.config.ProjectConfig;
 
 /**
  *
