+++ src/main/java/pseudopad/ui/FallbackPanel.java
@@ -14,7 +14,7 @@ import javax.swing.JPanel;
 import javax.swing.UIManager;
 import pseudopad.app.MainFrame;
 import pseudopad.ui.components.ImagePanel;
-import pseudopad.utils.AppActionsManager;
+import pseudopad.utils.ActionsManager;
 import pseudopad.utils.ThemeManager;
 
 /**
@@ -23,13 +23,13 @@ import pseudopad.utils.ThemeManager;
  */
 public class FallbackPanel extends JPanel {
     private final MainFrame appFrame;
-    
+
     public FallbackPanel(MainFrame appFrame) {
         this.appFrame = appFrame;
         updateState();
-        
+
         this.setLayout(new GridBagLayout());
-        
+
         if (appFrame.getCurrentProjectPath() == null) {
             initUIComponents();
 
@@ -46,89 +46,89 @@ public class FallbackPanel extends JPanel {
             this.add(new JLabel("Select a file to edit..."));
         }
     }
-    
+
     public final void updateState() {
         this.removeAll(); // Clear current view
-        
+
         if (appFrame.getCurrentProjectPath() == null) {
             // No Project -> Show Buttons
-            initUIComponents(); 
+            initUIComponents();
         } else {
             // Active Project -> Show "Select File" instruction
             this.setLayout(new GridBagLayout()); // Center it
             this.add(new JLabel("Select a file to edit..."));
         }
-        
+
         this.revalidate();
         this.repaint();
     }
-    
+
     private JButton createButton(String title, Action action) {
         JButton button = new JButton();
         button.setLayout(new FlowLayout(FlowLayout.LEFT, 24, 8));
-        
+
         button.addActionListener(action);
-        
+
         JPanel textContainer = new JPanel();
         textContainer.setOpaque(false);
         textContainer.setLayout(new BoxLayout(textContainer, BoxLayout.PAGE_AXIS));
-        
+
         JLabel titleLabel = new JLabel();
         titleLabel.setText(title);
-        
+
         JLabel shortcutLabel = new JLabel();
         shortcutLabel.setForeground(UIManager.getColor("Component.accentColor"));
-        
+
         if (action != null) {
             String actionShortcut = (String) action.getValue(Action.ACTION_COMMAND_KEY);
             shortcutLabel.setText(actionShortcut != null ? actionShortcut : "-");
         } else {
             shortcutLabel.setText("-");
         }
-        
+
         textContainer.add(titleLabel);
         textContainer.add(shortcutLabel);
-        
+
         button.add(textContainer);
-        
+
         button.putClientProperty(FlatClientProperties.BUTTON_TYPE, FlatClientProperties.BUTTON_TYPE_BORDERLESS);
-        
+
         return button;
     }
-    
+
     private void initUIComponents() {
         initComponents();
-        
+
         if (ThemeManager.getInstance().isDarkMode()) {
             brandImage.setImageResourcePath("/img/PseudoPad_Logomark_Dark.png");
         } else {
             brandImage.setImageResourcePath("/img/PseudoPad_Logomark.png");
         }
-        
+
         brandImage.setPreferredSize(new Dimension(150, 150));
-        
+
         body.add(brandImage, BorderLayout.WEST);
-        
+
         gridPanel.setLayout(new GridLayout(3, 0));
-        
-        AppActionsManager actionManager = appFrame.getAppActionInstance();
-        
+
+        ActionsManager actionManager = appFrame.getAppActionInstance();
+
         gridPanel.add(createButton("New Project...", actionManager.NEW_PROJECT));
         gridPanel.add(createButton("Open Project...", actionManager.OPEN_PROJECT));
         gridPanel.add(createButton("Open File...", null));
-        
+
         body.add(gridPanel, BorderLayout.EAST);
         body.setMinimumSize(new Dimension(350, 250));
-        
+
         this.add(body);
     }
-    
+
     private void initComponents() {
         this.body = new JPanel();
         this.brandImage = new ImagePanel();
         this.gridPanel = new JPanel();
     }
-    
+
     private JPanel body;
     private ImagePanel brandImage;
     private JPanel gridPanel;
