+++ src/main/java/pseudopad/ui/splash/SplashScreenPanel.form
