+++ src/main/java/pseudopad/ui/splash/SplashScreenPanel.java
@@ -1,4 +1,4 @@
-package pseudopad.ui;
+package pseudopad.ui.splash;
 
 import pseudopad.utils.ThemeManager;
 
@@ -13,29 +13,30 @@ public class SplashScreenPanel extends javax.swing.JPanel {
      */
     public SplashScreenPanel() {
         initComponents();
-        
+
         if (ThemeManager.getInstance().isDarkMode()) {
             ImagePanel.setImageResourcePath("/img/PseudoPad_SplashScreenTitle_Dark.png");
         } else {
             ImagePanel.setImageResourcePath("/img/PseudoPad_SplashScreenTitle.png");
         }
     }
-    
+
     public void setDebugText(String message) {
         SplashScreenText.setText(message);
     }
-    
+
     public void setProgressValue(int progress) {
         SplashScreenProgressBar.setValue(Math.max(0, Math.min(100, progress)));
     }
-    
+
     /**
      * This method is called from within the constructor to initialize the form.
      * WARNING: Do NOT modify this code. The content of this method is always
      * regenerated by the Form Editor.
      */
     @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    // <editor-fold defaultstate="collapsed" desc="Generated
+    // Code">//GEN-BEGIN:initComponents
     private void initComponents() {
 
         ImagePanel = new pseudopad.ui.components.ImagePanel();
@@ -59,13 +60,11 @@ public class SplashScreenPanel extends javax.swing.JPanel {
         javax.swing.GroupLayout ImagePanelLayout = new javax.swing.GroupLayout(ImagePanel);
         ImagePanel.setLayout(ImagePanelLayout);
         ImagePanelLayout.setHorizontalGroup(
-            ImagePanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 600, Short.MAX_VALUE)
-        );
+                ImagePanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGap(0, 600, Short.MAX_VALUE));
         ImagePanelLayout.setVerticalGroup(
-            ImagePanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 300, Short.MAX_VALUE)
-        );
+                ImagePanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGap(0, 300, Short.MAX_VALUE));
 
         add(ImagePanel, java.awt.BorderLayout.CENTER);
 
@@ -93,7 +92,6 @@ public class SplashScreenPanel extends javax.swing.JPanel {
 
         add(jPanel2, java.awt.BorderLayout.PAGE_END);
     }// </editor-fold>//GEN-END:initComponents
-
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private pseudopad.ui.components.ImagePanel ImagePanel;
